feat: Update README.md for Redmine 6.0.x.
Brief testing with Redmine 6.0.3 did not reveal any problems, except for
missing plugin assets. I had to manually copy the assets to the public
www root directory.
Unfortunately I could not find out from the documentation whether this
is expected or not. However, there was a change for Redmine 6.0.0 that
might be related: https://www.redmine.org/issues/39111